Move connection utility methods to SystemConnectionMemoManager
