+++ help/en/releasenotes/current-draft-note.shtml
@@ -580,9 +580,11 @@
     <h3>Miscellaneous</h3>
         <a id="Misc" name="Misc"></a>
         <ul>
-            <li><i>jmri.util.ConnectionUtil</i> has a utility method to find a connection by its connection name.<br>
-            <code>LocoNetSystemConnectionMemo memo = ConnectionUtil.getConnection("L2", LocoNetSystemConnectionMemo.class);</code><br>
-            <code>MqttSystemConnectionMemo memo = ConnectionUtil.getConnection("M", MqttSystemConnectionMemo.class);</code>
+            <li><i>jmri.jmrix.SystemConnectionMemoManager</i> has new utility methods to find a connection by its connection name.<br>
+            <code>LocoNetSystemConnectionMemo memo = SystemConnectionMemoManager.getConnection("L2", LocoNetSystemConnectionMemo.class);</code><br>
+            <code>LocoNetSystemConnectionMemo memo = SystemConnectionMemoManager.getConnectionByUserName("LocoNet", LocoNetSystemConnectionMemo.class);</code><br>
+            <code>MqttSystemConnectionMemo memo = SystemConnectionMemoManager.getConnection("M", MqttSystemConnectionMemo.class);</code>
+            <code>MqttSystemConnectionMemo memo = SystemConnectionMemoManager.getConnectionByUserName("MQTT", MqttSystemConnectionMemo.class);</code>
             </li>
         </ul>
